Travis/GH Actions: only run PHPCS once per build
All PHPCS sniffs are set up to provide the same results independently of the PHP version on which the sniff is run, so running PHPCS on every PHP version is unnecessary and highly inefficient.

This adds conditions to make sure the PHPCS action is only run on one PHP version.