Support IDN domain comparison
In order to determine the zone identifier, the plugin makes a call to
the Cloudflare API and fetchs all the zones the authorization method has
access to. From there, it compares the API response with the site
domain. Unfortunately, in the process of getting to the comparison the
value isn't always utf8 compatible resulting in a missed lookup even
though the zone is actually found in the API result as a utf8 value.

To rectify this, I've updated the comparisons to always convert to ASCII
before performing the comparison which ensures that if either WordPress
or the Cloudflare API respond with something in UTF8 or ASCII, we will
always find it correctly. I opted for ASCII comparisons (and storage)
here for two main reasons. The first is that not all database storage
systems available to WordPress are configured for UTF8. Attempting to
store and lookup these values may cause hard to debug issues. The second
is that ASCII is machine and human readable for comparisons, UTF8 can
have hidden characters or slight variances which the human eye will
miss.

Instead of relying on php-intl extensions for the IDN comparison, we are
using a symfony polyfill which ensures the best compatibility across
hosting providers.