rewrite symfony/polyfill bootstrap80.php contents instead of deleting
As a part of our deployment pipeline, we push the compiled plugin to
WordPress's SVN server. Unfortunately, the SVN precommit hook does not
yet support PHP 8[1] (which Symfony's intl/idn polyfill includes).
Despite WordPress only offering PHP 8 support in a beta[2] capacity[3],
people are running PHP 8 in production and the 4.5.0 release which
caused some sites unusable.

While this isn't ideal as we are rewriting the vendored package
contents, this does workaround the issue by removing the PHP 8 feature
that is causing the lint failure while still leaving the file in a
usable state.

[1]: https://meta.trac.wordpress.org/ticket/3791
[2]: https://make.wordpress.org/core/2020/11/23/wordpress-and-php-8-0/
[3]: https://make.wordpress.org/core/handbook/references/php-compatibility-and-wordpress-versions/