be smarter about cache purging for default types
By default, caching within Cloudflare is restricted to a handful of file
extensions[1]. However, within this WordPress plugin, we just pass it all the
cache purge irrespective of the file extension or zone configuration. This
means we're sending a boat load more URLs to the service than what we will
ever be purging.

To combat this we are introducing some additional logic that checks two things:

Firstly, we look at the page rules for the presence of any cache override
behaviour should it not be present, filter the URLs to only purge those in the
default caching extensions. This should be considered quite safe as out of the
box, everyone starts with this configuration and is only overridden when someone
explicitly sets this in page rules. We don't bother reimplementing the page rule
globbing logic but instead just check for the presence of "cache_everything"
for the cache level in any rule.

Secondly, we ensure the zone doesn't have APO enabled. APO is a feature that
relies on custom purge logic we don't want to duplicate here so just skip it if
it exists.

[1]: https://developers.cloudflare.com/cache/about/default-cache-behavior/#default-cached-file-extensions