Fix for APO not purging cache on page save
Saving changes on a page was not triggering cache purge for that page. Condition for testing used check for publicly_queryable prop, which by default return false for pages. However WP has build-in function for checking if queried object can be publicly viewed and returns correct values for all types of posts.