SPEED-777: Upgrade cloudflare-plugin-frontend to v3.8.0 (#534)
**🔖 Summary**

By upgrading the compiled files from `cloudflare-plugin-frontend` as of
v3.8.0, this commit aims at removing all Railgun-related code, which is
now a deprecated Cloudflare feature. This change is proposed due to a
`Bad Request` banner which is now shown on the Cloudflare Wordpress
plugin page because of a failing call to one of the deprecated Railgun
endpoints.

Close #534.