release 4.12.5
**🔖 Summary**

This release includes two minor fixes.

First one is making sure that the only the active zone is fetched when
retrieving the customer zone, preventing accidental fetching of a purged
zone with the same name, which happened to at least one customer.

Second change is about preventing some error banners from appearing on
the user plugin page, which was due to failing calls to retrieve details
about the Railgun settings of the user, which is now deprecated by
Cloudflare, and whose endpoints have been decomissioned.

**🚧 Changes**

- [CUSTESC-36595: Retrieve the Wordpress domain only from the active zone](https://github.com/cloudflare/Cloudflare-WordPress/pull/532)
- [SPEED-777: Upgrade cloudflare-plugin-frontend to v3.8.0](https://github.com/cloudflare/Cloudflare-WordPress/pull/535)