APO-286: Prevent default settings from enabling Auto Minify due to deprecation
This commit prevents the Auto Minify setting from being enabled whenever
a new user is enabling the default Cloudflare settings for Wordpress.

This is following the Auto Minify deprecation whose removal is scheduled
for August 5th, 2024.

Because the settings endpoint will remain available on Cloudflare API
and will eventually be made a no-op, the interface of the users of the
previous version of the Wordpress plugin will not see their interface break.

The `compiled.js` change, along with the `lang/*` changes were copied
after running `yarn build:production` following [this PR of the
`cloudflare-plugin-frontend` plugin](https://github.com/cloudflare/cloudflare-plugin-frontend/pull/202).